requirements.txt
streamlit==1.31.0
nltk==3.8.1
sentence-transformers==2.2.2
chromadb==0.4.18
openai==1.6.1
python-dotenv==1.0.0 